requirements.txt
fastapi
uvicorn
pydantic
google-genai
requests
python-multipart